pyproject.toml
[project]
name = "folio-updater"
version = "0.6.12"
authors = [{ name = "Deigue" }]
description = "Folio Updater"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.0.3",
    "pandas-market-calendars>=4.0.0",
    "fastparquet>=2024.11.0",
    "pyyaml>=6.0.2",
    "typer>=0.9.0",
    "keyring>=25.5.0",
    "ws-api>=0.22.0",
    "rich>=13.0.0",
]

[project.scripts]
folio = "cli.main:main"

[tool.pytest.ini_options]
# sync addopts with .vscode/settings.json
addopts = "--color=yes -vv"
log_cli = true
log_cli_level = "ERROR"
log_cli_format = "[%(levelname)s] %(name)s(%(lineno)d): %(message)s"
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "no_mock_forex: Disable the mock_forex_data fixture for this test or module.",
]

[tool.coverage.run]
branch = false
omit = [
    '*/__init__.py',
    'tests/fixtures/dataframe_cache.py',
    'src/utils/backup.py',
    'src/mock/folio_setup.py',
    'src/models/*',
    'src/cli/display.py',
    'src/cli/console.py'
]
[tool.coverage.report]
show_missing = true

exclude_lines = [
    "if TYPE_CHECKING:",
    "pragma: no cover",
    "except.*Error.*:",
    "except.*Exception.*:",
    "def __repr__",
    "def __str__",
    "def __len__",
    "if __name__ == .__main__.:",
    "def main",                   # CLI main function
    "typer.echo",                 # typer output
    "typer.Exit",                 # typer exit calls
    "console_error",              # ignore console error
    "console_warning",            # ignore console warning
    "logger.error",               # logging error handling
    "^\\s*\\w*msg\\s*=.*",        # generic log message assignments
    "^\\s*raise [A-Z]\\w*Error",  # raise errors
]


[tool.coverage.paths]

[tool.ruff]
line-length = 88
lint.extend-select = ["ALL"]
lint.ignore = [
    "D413",
    "INP001",
    "S608",
    "D203",
    "D213",
    "PGH004",
    "RUF001"
]
src = ["src"]
exclude = [
    "**/Lib/**",
    "**/coverage/**",
    "**/.venv/**",
    "**/.git/**",
    "**/.vscode/**",
]

[tool.ruff.lint.per-file-ignores]
"*.ipynb" = [
    "T201",
    "F811",
    "ERA001",
    "COM812",
]
"scripts/*" = ["T201"]
"tests/*" = [
    "S101",
    "D103",
    "PLR0913",
    "PLR2004",
    "T201",
]
"src/cli/main.py" = ["PLC0415"]
"src/cli/*" = [
    "PLR0913",
    "FBT003",
    "PLW0603",
    "ANN401",
]

[tool.ruff.lint.flake8-tidy-imports]
banned-module-level-imports = ["pandas_market_calendars"]

[tool.uv]
managed = true

#[tool.uv.sources]
#ws-api = { path = "../ws-api-python", editable = true }

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "nbstripout>=0.8.1",
    "pyinstaller>=6.16.0",
    "ruff>=0.12.10",
]
experiments = ["yfinance>=0.2.66"]
test = ["pytest>=8.3.5", "pytest-cov>=5.0.0"]
